requirements.txt
numpy
faiss-cpu
pandas
requests
ollama
sentence_transformers
msgpack